+++ railties/test/application/rake/dbs_test.rb
@@ -194,6 +194,16 @@ module ApplicationTests
         db_structure_dump_and_load database_url_db_name
       end
 
+      test "db:structure:dump and db:structure:load set ar_internal_metadata" do
+        require "#{app_path}/config/environment"
+        db_structure_dump_and_load ActiveRecord::Base.configurations[Rails.env]["database"]
+
+        Dir.chdir(app_path) do
+          assert_equal "test", `bin/rails runner -e test "puts ActiveRecord::InternalMetadata[:environment]"`.strip
+          assert_equal "development", `bin/rails runner "puts ActiveRecord::InternalMetadata[:environment]"`.strip
+        end
+      end
+
       test "db:structure:dump does not dump schema information when no migrations are used" do
         Dir.chdir(app_path) do
           # create table without migrations
@@ -304,6 +314,49 @@ module ApplicationTests
         ensure
           ENV["RAILS_ENV"] = @old_rails_env
           ENV["RACK_ENV"] = @old_rack_env
+        end
+      end
+
+      test "db:setup sets ar_internal_metadata" do
+        Dir.chdir(app_path) do
+          app_file "db/schema.rb", ""
+          `bin/rails db:setup`
+
+          test_environment = lambda { `bin/rails runner -e test "puts ActiveRecord::InternalMetadata[:environment]"`.strip }
+          development_environment = lambda { `bin/rails runner "puts ActiveRecord::InternalMetadata[:environment]"`.strip }
+
+          assert_equal "test", test_environment.call
+          assert_equal "development", development_environment.call
+
+          app_file "db/structure.sql", ""
+          app_file "config/initializers/enable_sql_schema_format.rb", <<-RUBY
+            Rails.application.config.active_record.schema_format = :sql
+          RUBY
+
+          `bin/rails db:setup`
+
+          assert_equal "test", test_environment.call
+          assert_equal "development", development_environment.call
+        end
+      end
+
+      test "db:test:prepare sets test ar_internal_metadata" do
+        Dir.chdir(app_path) do
+          app_file "db/schema.rb", ""
+          `bin/rails db:test:prepare`
+
+          test_environment = lambda { `bin/rails runner -e test "puts ActiveRecord::InternalMetadata[:environment]"`.strip }
+
+          assert_equal "test", test_environment.call
+
+          app_file "db/structure.sql", ""
+          app_file "config/initializers/enable_sql_schema_format.rb", <<-RUBY
+            Rails.application.config.active_record.schema_format = :sql
+          RUBY
+
+          `bin/rails db:test:prepare`
+
+          assert_equal "test", test_environment.call
         end
       end
     end